+++ src/web/twig/variables/VideosVariable.php
@@ -24,17 +24,18 @@ class VideosVariable
      * Get embed html from a video url.
      *
      * @param string $videoUrl
-     * @param array $options
+     * @param array $htmlOptions
+     * @param array $urlOptions
      * @return null|string
      *
      * @since 3.0.0
      */
-    public function getEmbedHtml(string $videoUrl, array $options = []): ?string
+    public function getEmbedHtml(string $videoUrl, array $htmlOptions = [], array $urlOptions = []): ?string
     {
         try {
             $video = VideosPlugin::$plugin->getVideos()->getVideoByUrl($videoUrl);
 
-            return $video->getEmbedHtml($options);
+            return $video->getEmbedHtml($htmlOptions, $urlOptions);
         } catch (Exception $e) {
             Craft::info('Couldn’t get video from its url ('.$videoUrl.'): '.$e->getMessage(), __METHOD__);
         }