+++ CHANGELOG.md
@@ -85,8 +85,8 @@ Changelog
 - Changed `dukt\videos\base\Gateway::getVideoByUrl()` is now final
 - Changed `dukt\videos\base\Gateway::getVideoById()` is now final ; use `dukt\videos\base\Gateway::fetchVideoById()` to get video from the gateway API
 - Moved `dukt\videos\base\Gateway::getEmbedFormat()` to `dukt\videos\base\Gateway::getEmbedUrlFormat()`
-- Moved `dukt\videos\base\Gateway::getEmbed()` to `dukt\videos\helpers\EmbedHelper::getEmbedHtml()` and need `dukt\videos\model\Video` as params (instead of video id)
-- Moved `dukt\videos\base\Gateway::getEmbedUrl()`  to `dukt\videos\helpers\EmbedHelper::getEmbedUrl()` and need `dukt\videos\model\Video` as params (instead of video id)
+- Moved `dukt\videos\base\Gateway::getEmbed()` to `dukt\videos\helpers\EmbedHelper::getEmbedHtml()` and need `dukt\videos\model\Video` as param (instead of video id) and there is a split of html/url options in params
+- Moved `dukt\videos\base\Gateway::getEmbedUrl()`  to `dukt\videos\helpers\EmbedHelper::getEmbedUrl()` and need `dukt\videos\model\Video` as param (instead of video id)
 - Changed `dukt\videos\base\Gateway::getVideos()` is now final
 - Moved `dukt\videos\base\Gateway::getExplorerSections()` to `dukt\videos\base\Gateway::getExplorer()`
 - Moved `dukt\videos\base\Gateway::extractVideoIdFromUrl()` to `dukt\videos\base\Gateway::extractVideoIdFromVideoUrl()`